Support configurability.
Build arbitrary stages or choose from precanned db, agent, or master
stages.

Just embed the master and agent configs right in the yaml file, to make
it as flexible as possible (and to offload as much logic as possible
onto the application rather than the config parsing).

Hasura is supposed to go away soon, so it has to be configured as a
totally custom stage.