Various improvements.
Use pipe() instead of pipe2().  Sorry, macOS, I don't know why I thought
you could do this.

Support `/~ as alternative keys to 0/).

Support a new "startup_input" top-level key of .devcluster.yaml.  The
characters of startup_input act as if you typed them into the terminal
UI as soon as the UI was up.

Add the "logcheck" AtomicOperation that can match lines of output
against a regex pattern to identify when a stage is up.