Support generic builds as pre commands.
Build commands change from time to time, so they should be part of the
config file.

Support configurale binary paths for master and agent, so that the
output directory is allowed to change.

Add a new sh option as a pre command which is just a nice shortcut for
defining custom commands via shell.