Run subprocesses as separate process groups.
It was not safe to send a signal to the devcluster process, because some
child processes would also receive the signal, but the docker container
would not.  As a result, after a SIGNIT, devcluster would see
`docker container logs` exit and would try a `docker container wait`,
which would always hang.

An alternate strategy would have been to manually forward the signal to
the docker container a la `docker kill --signal=INT`.  I'm not sure I
fully understand the tradeoff between the two strategies.