Support multiple agents in a single devcluster.
You can provide a unique `name` field to every agent to deconflict the
agent config files, log files, and log stream names.

This is useful for testing dtrain scheduling behavior, for instance.

It is not useful for testing dtrain harness behavior because our harness
will not behave itself with multiple agents on a single machine.